Track explicitly when the bar is ready to use
This should prevent any instances of trying to update the bar when it's
not ready. I'm not certain exactly how this could happen, but my best
guess is that it could be cased by a call to Update() at the instant
that the bar is being rebuilt after a settings change.